have to put it outside hazard model to avoid circular dependency
