+++ pyincore/models/hazard/earthquake.py
@@ -3,7 +3,7 @@
 # This program and the accompanying materials are made available under the
 # terms of the Mozilla Public License v2.0 which accompanies this distribution,
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
-from pyincore import HazardService, Dataset
+from pyincore import HazardService
 from pyincore.models.hazard.hazard import Hazard
 from pyincore.models.hazard.hazarddataset import EarthquakeDataset
